feat: 拆分 JSON 语法高亮适配器
