+++ Sources/Lithe/Views/Editor/SyntaxHighlighting/JSONSyntaxHighlightingAdapter.swift
@@ -1,0 +1,17 @@
+import AppKit
+
+/// Adds JSON object-property treatment after generic string tokenization.
+enum JSONSyntaxHighlightingAdapter {
+    private static let propertyExpression = try! NSRegularExpression(
+        pattern: #"\"(?:\\.|[^\"\\])*\"(?=\s*:)"#
+    )
+
+    static func apply(to storage: NSTextStorage, palette: CodeEditorPalette, range: NSRange) {
+        GenericSyntaxHighlightingAdapter.apply(
+            propertyExpression,
+            color: palette.property,
+            storage: storage,
+            range: range
+        )
+    }
+}
