+++ Tests/LitheTests/Fixtures/SyntaxHighlighting/json-syntax-corpus.json
@@ -1,0 +1,18 @@
+{
+  "name": "Lithe",
+  "number": 42,
+  "negativeDecimal": -3.14,
+  "scientific": 6.02e23,
+  "boolean": true,
+  "nullValue": null,
+  "escaped\"key": "escaped\\value",
+  "nested": {
+    "items": [
+      "name",
+      {
+        "name": "nested-value"
+      }
+    ]
+  },
+  "tail": ["name"]
+}
